--- a/website/thaliawebsite/settings.py
+++ b/website/thaliawebsite/settings.py
@@ -468,6 +468,7 @@ if DATABASE_ENGINE == "postgresql":
             "NAME": os.environ.get("POSTGRES_DB", ""),
             "HOST": os.environ.get("POSTGRES_HOST", ""),
             "PORT": os.environ.get("POSTGRES_PORT", "5432"),
+            "CONN_MAX_AGE": 300,
         }
     }