@DeD1rk 2023-12-24
```
# With CONN_MAX_AGE
❯ hey -c 20 -n 400 https://localhost

Summary:
  Total:	9.6618 secs
  Slowest:	6.7543 secs
  Fastest:	0.0337 secs
  Average:	0.4094 secs
  Requests/sec:	41.4002
  

Response time histogram:
  0.034 [1]	|
  0.706 [365]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.378 [14]	|■■
  2.050 [8]	|■
  2.722 [3]	|
  3.394 [1]	|
  4.066 [0]	|
  4.738 [0]	|
  5.410 [0]	|
  6.082 [1]	|
  6.754 [7]	|■


Latency distribution:
  10% in 0.0952 secs
  25% in 0.1234 secs
  50% in 0.1959 secs
  75% in 0.2925 secs
  90% in 0.5872 secs
  95% in 1.4317 secs
  99% in 6.4647 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0149 secs, 0.0337 secs, 6.7543 secs
  DNS-lookup:	0.0021 secs, 0.0000 secs, 0.0477 secs
  req write:	0.0000 secs, 0.0000 secs, 0.0020 secs
  resp wait:	0.3936 secs, 0.0335 secs, 6.4647 secs
  resp read:	0.0003 secs, 0.0002 secs, 0.0024 secs

Status code distribution:
  [200]	400 responses

# Without CONN_MAX_AGE
❯ hey -c 20 -n 400 https://localhost

Summary:
  Total:	36.5971 secs
  Slowest:	9.3495 secs
  Fastest:	0.1260 secs
  Average:	1.6811 secs
  Requests/sec:	10.9298
  

Response time histogram:
  0.126 [1]	|
  1.048 [198]	|■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  1.971 [76]	|■■■■■■■■■■■■■■■
  2.893 [59]	|■■■■■■■■■■■■
  3.815 [26]	|■■■■■
  4.738 [21]	|■■■■
  5.660 [5]	|■
  6.582 [5]	|■
  7.505 [1]	|
  8.427 [4]	|■
  9.349 [4]	|■


Latency distribution:
  10% in 0.3111 secs
  25% in 0.5263 secs
  50% in 1.0713 secs
  75% in 2.3811 secs
  90% in 3.8848 secs
  95% in 4.6989 secs
  99% in 8.5517 secs

Details (average, fastest, slowest):
  DNS+dialup:	0.0076 secs, 0.1260 secs, 9.3495 secs
  DNS-lookup:	0.0010 secs, 0.0000 secs, 0.0296 secs
  req write:	0.0001 secs, 0.0000 secs, 0.0075 secs
  resp wait:	1.6725 secs, 0.1258 secs, 9.0782 secs
  resp read:	0.0004 secs, 0.0001 secs, 0.0062 secs

Status code distribution:
  [200]	400 responses
```